persist user across app using Provider/UserContext
