+++ front-end/src/components/ConnectWalletButton.js
@@ -1,39 +1,15 @@
 import {Button, Box, Text, Center } from "@chakra-ui/react"
 import {ethers} from 'ethers';
+import { useState, useContext } from "react";
+import { getAddress } from "../lib/metamask";
+import { UserContext } from "../lib/UserContext";
+
 export default function ConnectWalletButton ({ onClick }) {
+    const { user, setUser } = useContext(UserContext);
 
-    async function connectWallet() {
-        const provider = new ethers.providers.Web3Provider(window.ethereum, "rinkeby");
-        // Check if MetaMask installed
-        if (window.ethereum) {
-            // check if correct network connected
-            const chainId = await window.ethereum.request({ method: 'eth_chainId'});
-
-            console.log(chainId)
-            if ( chainId != "0x4") {
-                alert("Please change to Rinkeby/")
-            }
-            else {
-                
-                await provider.send("eth_requestAccounts", []);
-                const signer = provider.getSigner();
-
-                if (onClick !== undefined) {
-                    onClick(await signer.getAddress());
-                }
-
-                console.log("Account:", await signer.getAddress());
-                const balance = await provider.getBalance(signer.getAddress());
-                console.log("Balance: ", balance);
-            }
-        }
-        else{
-            alert("No wallet detected, please install metamask.")
-        }
-    }
     return (<Center>
-        <Button mt={4} colorScheme='teal' type='submit' className="connectButton" onClick={connectWallet} >
-        Connect Wallet
+        <Button colorScheme='teal' type='submit' className="connectButton" onClick={async () => await getAddress()} >
+            {user.address? 'Wallet Connected': 'Connect Wallet'}
         </Button>
         </Center>
     )
