+++ front-end/src/lib/UserContext.js
@@ -1,4 +1,5 @@
-import { createContext, useMemo, useState } from "react";
+import { createContext, useMemo, useState, useEffect } from "react";
+import { getAddress } from "./metamask";
 
 export const UserContext = createContext({});
 
@@ -6,6 +7,14 @@ export function UserContextProvider({ children }) {
 
   const [user, setUser] = useState({});
   const value = useMemo(() => ({ user, setUser }), [user, setUser]);
+
+  useEffect(() => {
+    const setAddress = async () => {
+      setUser({ address: await getAddress() })
+    }
+    setAddress();
+    console.log("User was fetched")
+}, []);
 
   return (
     <UserContext.Provider value={value}>
