+++ front-end/src/pages/TasksPage/index.js
@@ -18,7 +18,6 @@ import {
   import Header from '../../components/Header'
 import ConnectWalletButton from '../../components/ConnectWalletButton';
 import { UserContext } from '../../lib/UserContext';
-import { getAddress } from '../../lib/metamask';
 
   const ACTIVE = "active"
   const EXPIRED = "expired"
@@ -189,15 +188,12 @@ import { getAddress } from '../../lib/metamask';
 
   export function TasksPage() {
     const [data, setData] = useState(testData);
+    const { user, setUser } = useContext(UserContext);
 
     useEffect(() => {
+        // console.log("User fetched", user)
         const url = "http://localhost:3042/tasks?showCompleted=true";
         console.log("Fetched")
-
-        const setAddress = async () => {
-          setUser({ address: await getAddress() })
-        }
-        setAddress();
 
         const fetchData = async () => {
             try {
@@ -229,11 +225,8 @@ import { getAddress } from '../../lib/metamask';
         fetchData();
     }, []);
 
-    const { user, setUser } = useContext(UserContext);
-
     return (
       <Box maxW="7xl" mx={'auto'} pt={5} px={{ base: 2, sm: 12, md: 17 }}>
-        <ConnectWalletButton onClick={address => setUser({ address })}/>
         <chakra.h1
           textAlign={'center'}
           fontSize={'4xl'}
